Link to correctly functioning express-route-tester
The version that was linked to uses a different version of path-to-regex
so produces results that are not useful for writing express routes.

This version lets you select the path-to-regex version you want to use
and defaults to the version used by the current version of express.

It also highlights when a regex is invalid, whereas the existing one
will silently carry on working with the last-known-good regex.

I've also the URL to use https for best-practise reasons.